Update for 4.4 with SELinux enabled condition

+++ scripts/build.sh
@@ -6,8 +6,12 @@ cd work
 
 # change version here
 wget -c $ROM_URL -O dl_rom.zip
-unzip dl_rom.zip
-unzip $ZIP_FILE -d unzipped_rom
+if [ -n "$ZIP_FILE" ]; then
+  unzip dl_rom.zip
+  unzip $ZIP_FILE -d unzipped_rom
+else
+  unzip dl_rom.zip -d unzipped_rom
+fi
 
 # just to get file_contexts
 #./unpack_intel UL-ASUS_T00F-WW-1.17.40.16-user/boot.img bzImage ramdisk.cpio.gz
@@ -16,7 +20,9 @@ unzip $ZIP_FILE -d unzipped_rom
 #cd ..
 
 mv unzipped_rom/system .
-rm -rf $ZIP_FILE
+if [ -f $ZIP_FILE ]; then
+  rm -rf $ZIP_FILE
+fi
 rm -rf unzipped_rom
 
 cp -R ../root/* system
